Separated JWT creation to different folder
